made a basic example of rotating boxes
