+++ projects/oscillation/rotation/main.py
@@ -1,5 +1,5 @@
 import sys
-sys.path.append("../../..lib")
+sys.path.append("../../../lib")
 
 from sketchy import Sketchy
 from vector import Vector
@@ -10,9 +10,9 @@ class MyDrawing(Sketchy):
     def setup(self):
         width, height = [600, 400]
         self.balls = []
-        for i in range(0, 1):
+        for i in range(0, 20):
             self.balls.append(Ball(width,height))
-            self.size(width, height)
+        self.size(width, height)
 
     def update(self):
         for i in range(len(self.balls)):
@@ -20,7 +20,9 @@ class MyDrawing(Sketchy):
 
     def draw(self, g):
         g.background(1, 1, 1)
+        # g.push()
         for i in range(len(self.balls)):
-            self.balls[i].draw(g)
+            self.balls[i].draw(g, 600, 400)
+        # g.pop()
 
 MyDrawing()
